+++ blocks/ai-image/src/icons/abstract-ai.tsx
